+++ ChangeLog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ## Version 1.14
 
+- FIX : Sprint and Kanban times updates *09/11/2022* - 1.14.3
 - FIX : Assign user to right element or card and update context menu *04/11/2022* - 1.14.2
 - FIX : Contact tab display for scrum task and scrum Us *26/10/2022* - 1.14.1
 - NEW : Option for kanban status display *26/10/2022* - 1.14.0